tests: Extract TestOptions into its own file
